Add admin tools in views + merge backend with frontend part in 1 entity
