+++ index.php
@@ -1,9 +1,65 @@
 <?php
 
-require('controller/frontend.php');
+require('controller.php');
 
 try
 {
+	//BACK
+	if (isset($_POST['title']) AND isset($_POST['post_content']))
+	{
+		if (! empty($_POST['title']) AND ! empty($_POST['post_content']))
+		{
+			adminSendPost();
+		}
+		else
+		{
+			header('Location: ../view/backend/add_form.php?empty_field=true');
+		}
+	}
+	
+	if(isset($_GET['send_post']) AND ($_GET['send_post']))
+	{
+		echo '<p>Le billet a bien été ajouté !</p>';
+	}
+	
+	
+	if (isset($_GET['delete_post']))
+	{
+		adminDeletePost();
+	}
+	
+	if(isset($_GET['deleted_post']) AND ($_GET['deleted_post']))
+	{
+		echo '<p>Le billet a bien été supprimé !</p>';
+	}
+	
+	
+	if(isset($_GET['edit_post']))
+	{
+		formEditPost();
+	}
+	
+	if(isset($_GET['sent_edit_post']))
+	{
+		adminEditPost();
+	}
+	
+	if(isset($_GET['edited_post']) AND ($_GET['edited_post']))
+	{
+		echo '<p>Le billet a bien été modifié !</p>';
+	}
+	
+	
+	if (isset($_GET['delete_comment']))
+	{
+		adminDeleteComment();
+	}
+	
+	if(isset($_GET['deleted_comment']) AND ($_GET['deleted_comment']))
+	{
+		echo '<p>Le commentaire a bien été supprimé !</p>';
+	}
+	//FRONT
     if (! isset($_GET['page']) OR $_GET['page'] > 1000 OR $_GET['page'] < 0)
     {
         $_GET['page'] = 1;
@@ -33,7 +89,6 @@ try
             header('Location: index.php?comment=' . $_GET['send_comment'] . '&empty=true');
         }
     }
-
 
     if (! isset($_GET['comment']) OR $_GET['comment'] > 1000 OR $_GET['comment'] <= 0)
     {
@@ -98,6 +153,8 @@ try
 			header('Location: view/frontend/connexion.php?connect=false');
 		}
 	}
+	
+
 }
 
 catch(Exception $e)
