+++ model/PagesManager.php
