--- view/backend/.htaccess
@@ -1,5 +1,0 @@
-AuthName "Page d'administration protégée"
-AuthType Basic
-AuthUserFile "C:\wamp64\www\blog_with_comments\view\backend\.htpasswd"
-Require valid-user
-RewriteEngine on
