--- view/backend/.htpasswd
@@ -1,1 +1,0 @@
-drx85:kangourou
