+++ view/template.php
@@ -5,7 +5,6 @@
 <head>
     <meta charset="UTF-8">
     <title>Bienvenue sur mon blog !</title>
-
 </head>
 
 <body>
@@ -21,11 +20,9 @@ if (! isset($_SESSION['user_id'])) {
 	<a href = "/p5_personnal_blog/index.php?p=connexion"> Se connecter </a>';
 }
 else {
-	echo 'Connecté en tant que ' . $_SESSION['pseudo'] . '</br>
+	echo 'Connecté en tant que ' . $_SESSION['pseudo'] . ' (' . $_SESSION['role'] . ')</br>
 	<a href="/p5_personnal_blog/view/frontend/disconnection.php?p=connexion">Se déconnecter</a>';
 }
-
-echo $_SESSION['role'];
 
 ?>
